The following update is to address the use of unicode throghout the pipeline. In the future, it is probably better to adopt Python3 from the very beginning.
StandardFileExtract worker will explicitly open XML files in UTF-8 with the codecs module.
Text files are attempted to be decoded once opened (this could be changed in the future).
Once opened, everything should remain in unicode (this is where problems can originate).
On output, WriteMetaFile explicitly writes fulltext content as bytes in UTF-8.
WriteMetaFile explicitly writes meta.json files in ASCII via the JSON library, which
retains unicode numbers to be read back in with the JSON library.

Within Java, everything is treated as Unicode without any intervention. The bytes
sent to RabbitMQ are now explicitly set to be UTF-8 encoded. Text is now also
normalised in the same way that they are in the Python module.

Tests have been updated to reflect the updates to the Java pipeline. Also, a new
PDF that includes Unicode characters has been added.